+++ js/mod.js
@@ -9,13 +9,13 @@ let modInfo = {
     offlineLimit: 1,  // In hours
     initialStartPoints: new Decimal(10), // Used for hard resets and new players
 	endgame: new Decimal("1e9999999999"),
-	specialEndgameText: 'v0.4 Beta 6 Endgame: e16,875 Points, 1e150 Quirk Energy',
+	specialEndgameText: 'v0.4 Beta 7 Endgame: e23,450 Points',
 }
 
 // Set your version in num and name
 let VERSION = {
 	num: "0.4",
-	beta: 6,
+	beta: 7,
 	name: "Hindered Quirkiness",
 }